moved services/ to lib/
