+++ src/lib/bootstrap/index.ts
@@ -1,0 +1,5 @@
+import initFirebase from './firebase'
+
+export default function bootstrap() {
+  initFirebase()
+}
