+++ src/pages/communities/create.tsx
@@ -1,4 +1,3 @@
-
 import { useRouter } from 'next/router'
 
 import CommunityCreate from '../../components/organisms/CommunityCreate'
